updates activeCluster on auth service and adds activeSession prop for sidebar visibility
