+++ ui/app/components/cluster-info.js
@@ -17,10 +17,6 @@ export default class ClusterInfoComponent extends Component {
   @service store;
   @service version;
 
-  get activeCluster() {
-    return this.store.peekRecord('cluster', this.auth.activeCluster);
-  }
-
   transitionToRoute() {
     this.router.transitionTo(...arguments);
   }
