+++ ui/app/controllers/vault/cluster.js
@@ -2,7 +2,7 @@
 import { inject as service } from '@ember/service';
 import { alias } from '@ember/object/computed';
 import Controller from '@ember/controller';
-import { observer, computed } from '@ember/object';
+import { observer } from '@ember/object';
 export default Controller.extend({
   auth: service(),
   store: service(),
@@ -31,35 +31,7 @@ export default Controller.extend({
   }),
 
   consoleOpen: alias('console.isOpen'),
-
-  activeCluster: computed('auth.activeCluster', function () {
-    return this.store.peekRecord('cluster', this.auth.activeCluster);
-  }),
-
-  activeClusterName: computed('activeCluster', function () {
-    const activeCluster = this.activeCluster;
-    return activeCluster ? activeCluster.get('name') : null;
-  }),
-
-  showNav: computed(
-    'router.currentRouteName',
-    'activeClusterName',
-    'auth.currentToken',
-    'activeCluster.{dr.isSecondary,needsInit,sealed}',
-    function () {
-      if (this.activeCluster.dr?.isSecondary || this.activeCluster.needsInit || this.activeCluster.sealed) {
-        return false;
-      }
-      if (
-        this.activeClusterName &&
-        this.auth.currentToken &&
-        this.router.currentRouteName !== 'vault.cluster.auth'
-      ) {
-        return true;
-      }
-      return;
-    }
-  ),
+  activeCluster: alias('auth.activeCluster'),
 
   actions: {
     toggleConsole() {
