+++ ui/app/templates/components/cluster-info.hbs
@@ -1,7 +1,7 @@
 <div class="popup-menu-content">
   <div class="box">
     {{#unless this.version.isOSS}}
-      {{#if (and this.activeCluster.unsealed this.auth.currentToken)}}
+      {{#if (and this.auth.activeCluster.unsealed this.auth.currentToken)}}
         {{#if @cluster.dr.isSecondary}}
           {{#if (has-permission "status" routeParams="replication")}}
             <nav class="menu" aria-label="replication secondary menu">
@@ -75,7 +75,7 @@
       </div>
       <ul class="menu-list">
         <li class="action">
-          {{#if this.activeCluster.unsealed}}
+          {{#if this.auth.activeCluster.unsealed}}
             {{#if (and (has-permission "status" routeParams="seal") (not @cluster.dr.isSecondary))}}
               <LinkTo @route="vault.cluster.settings.seal" @model={{@cluster.name}} {{on "click" @onLinkClick}}>
                 <div class="level is-mobile">
@@ -113,7 +113,7 @@
         <ul class="menu-list">
           {{#if (and this.version.features (has-permission "status" routeParams="license") (not @cluster.dr.isSecondary))}}
             <li class="action">
-              <LinkTo @route="vault.cluster.license" @model={{this.activeCluster.name}} {{on "click" @onLinkClick}}>
+              <LinkTo @route="vault.cluster.license" @model={{this.auth.activeCluster.name}} {{on "click" @onLinkClick}}>
                 <div class="level is-mobile">
                   <span class="level-left">License</span>
                   <Chevron class="has-text-grey-light level-right" />
@@ -123,7 +123,7 @@
           {{/if}}
           {{#if (and @cluster.usingRaft (has-permission "status" routeParams="raft"))}}
             <li class="action">
-              <LinkTo @route="vault.cluster.storage" @model={{this.activeCluster.name}} {{on "click" @onLinkClick}}>
+              <LinkTo @route="vault.cluster.storage" @model={{this.auth.activeCluster.name}} {{on "click" @onLinkClick}}>
                 <div class="level is-mobile">
                   <span class="level-left">Raft Storage</span>
                   <Chevron class="has-text-grey-light level-right" />
