+++ ui/app/templates/components/splash-page.hbs
@@ -6,7 +6,7 @@
       </HomeLink>
     </Nav.home>
     <Nav.items>
-      <div class="navbar-item status-indicator-button" data-status={{if this.activeCluster.unsealed "good" "bad"}}>
+      <div class="navbar-item status-indicator-button" data-status={{if this.auth.activeCluster.unsealed "good" "bad"}}>
         <StatusMenu @label="Status" @onLinkClick={{Nav.closeDrawer}} />
       </div>
     </Nav.items>
